+++ src-docs/src/views/button/button_example.js
@@ -14,6 +14,7 @@ import {
   EuiButtonGroup,
   EuiButtonToggle,
   EuiCallOut,
+  EuiText,
 } from '../../../../src/components';
 
 import Button from './button';
@@ -108,6 +109,16 @@ const buttonGroupSnippet = [
 
 export const ButtonExample = {
   title: 'Button',
+  intro: (
+    <EuiText>
+      <p>
+        <strong>EuiButton</strong> comes in two styles. The{' '}
+        <EuiCode>fill</EuiCode> style should be reserved for the main action and
+        limited in number for a single page. Be sure to read the full{' '}
+        <Link to="/guidelines/buttons">button usage guidelines</Link>.
+      </p>
+    </EuiText>
+  ),
   sections: [
     {
       source: [
@@ -146,7 +157,7 @@ export const ButtonExample = {
       demo: <ButtonAsLink />,
     },
     {
-      title: 'Button with Icon',
+      title: 'Button with icon',
       source: [
         {
           type: GuideSectionTypes.JS,
@@ -159,13 +170,16 @@ export const ButtonExample = {
       ],
       text: (
         <p>
-          The passed icon needs to come from our list of svg icons. Can be
-          flipped{' '}
+          The passed <EuiCode>iconType</EuiCode> must be an acceptable{' '}
+          <Link to="/display/icons">
+            <strong>EuiIcon</strong>
+          </Link>{' '}
+          type. It can be flipped{' '}
           {
             // eslint-disable-next-line react/no-unescaped-entities
           }{' '}
-          to the other side by passing{' '}
-          <EuiCode>iconSide=&quot;right&quot;</EuiCode>.
+          to the opposite side by passing{' '}
+          <EuiCode language="js">iconSide=&quot;right&quot;</EuiCode>.
         </p>
       ),
       snippet: buttonWithIconSnippet,
@@ -195,7 +209,7 @@ export const ButtonExample = {
       demo: <ButtonLoading />,
     },
     {
-      title: 'ButtonEmpty',
+      title: 'Empty button',
       source: [
         {
           type: GuideSectionTypes.JS,
@@ -208,9 +222,9 @@ export const ButtonExample = {
       ],
       text: (
         <p>
-          <EuiCode>EuiButtonEmpty</EuiCode> is used when you want to make a
-          button look like a regular link, but still want to align it to the
-          rest of the buttons.
+          <strong>EuiButtonEmpty</strong> is used when you want to make a button
+          look like a regular link, but still want to align it to the rest of
+          the buttons.
         </p>
       ),
       props: { EuiButtonEmpty },
@@ -218,7 +232,7 @@ export const ButtonExample = {
       demo: <ButtonOption />,
     },
     {
-      title: 'Flush ButtonEmpty',
+      title: 'Flush empty button',
       source: [
         {
           type: GuideSectionTypes.JS,
@@ -231,7 +245,7 @@ export const ButtonExample = {
       ],
       text: (
         <p>
-          When aligning <EuiCode>EuiButtonEmpty</EuiCode> components to the left
+          When aligning <strong>EuiButtonEmpty</strong> components to the left
           or the right, you should make sure they&rsquo;re flush with the edge
           of their container, so that they&rsquo;re horizontally aligned with
           the other content in the container.
@@ -241,7 +255,7 @@ export const ButtonExample = {
       demo: <ButtonOptionFlush />,
     },
     {
-      title: 'Button Icon',
+      title: 'Button icon',
       source: [
         {
           type: GuideSectionTypes.JS,
@@ -253,7 +267,10 @@ export const ButtonExample = {
         },
       ],
       text: (
-        <p>Button icons are buttons that only contain an icon (no text).</p>
+        <p>
+          <strong>EuiButtonIcons</strong> are buttons that only contain an icon
+          (no text).
+        </p>
       ),
       props: { EuiButtonIcon },
       snippet: buttonIconSnippet,
@@ -275,15 +292,15 @@ export const ButtonExample = {
         <div>
           <p>
             This is a specialized component that combines{' '}
-            <EuiCode>EuiButton</EuiCode> and <EuiCode>EuiToggle</EuiCode> to
-            create a button with an on/off state. You can pass all the same
-            parameters to it as you can to <EuiCode>EuiButton</EuiCode>. The
-            main difference is that, it does not accept any children, but a{' '}
+            <strong>EuiButton</strong> and <strong>EuiToggle</strong> to create
+            a button with an on/off state. You can pass all the same parameters
+            to it as you can to <strong>EuiButton</strong>. The main difference
+            is that, it does not accept any children, but a{' '}
             <EuiCode>label</EuiCode> prop instead. This is for the handling of
-            accessibility with the <EuiCode>EuiToggle</EuiCode>.
+            accessibility with the <strong>EuiToggle</strong>.
           </p>
           <p>
-            The <EuiCode>EuiButtonToggle</EuiCode> does not have any inherit
+            The <strong>EuiButtonToggle</strong> does not have any inherit
             visual state differences. These you must apply in your
             implementation.
           </p>
@@ -294,7 +311,7 @@ export const ButtonExample = {
       props: { EuiButtonToggle },
     },
     {
-      title: 'Groups',
+      title: 'Button groups',
       source: [
         {
           type: GuideSectionTypes.JS,
@@ -308,19 +325,19 @@ export const ButtonExample = {
       text: (
         <div>
           <p>
-            Button groups are handled similarly to the way checkbox and radio
-            groups are handled but made to look like buttons. They group
-            multiple <EuiCode>EuiButtonToggle</EuiCode>s and utilize the{' '}
-            <EuiCode>type=&quot;single&quot;</EuiCode> or{' '}
-            <EuiCode>&quot;multi&quot;</EuiCode> prop to determine whether
-            multiple or only single selections are allowed per group.
+            <strong>EuiButtonGroups</strong> are handled similarly to the way
+            checkbox and radio groups are handled but made to look like buttons.
+            They group multiple <strong>EuiButtonToggles</strong> and utilize
+            the <EuiCode language="js">type=&quot;single&quot;</EuiCode> or{' '}
+            <EuiCode language="js">&quot;multi&quot;</EuiCode> prop to determine
+            whether multiple or only single selections are allowed per group.
           </p>
           <p>
             Stylistically, all button groups are the size of small buttons, do
             not stretch to fill the container, and typically should only be{' '}
-            <EuiCode>color=&quot;text&quot;</EuiCode> (default) or{' '}
-            <EuiCode>&quot;primary&quot;</EuiCode>. If you&apos;re just
-            displaying a group of icons, add the prop{' '}
+            <EuiCode language="js">color=&quot;text&quot;</EuiCode> (default) or{' '}
+            <EuiCode language="js">&quot;primary&quot;</EuiCode>. If you&apos;re
+            just displaying a group of icons, add the prop{' '}
             <EuiCode>isIconOnly</EuiCode>.
           </p>
           <EuiCallOut
@@ -355,11 +372,14 @@ export const ButtonExample = {
       text: (
         <p>
           For buttons on dark color backgrounds, you can pass{' '}
-          <EuiCode>color=&apos;ghost&apos;</EuiCode> to any of the button styles
-          on this page. These should be used extremely rarely, and are only for
-          placing buttons on top of dark or image-based backgrounds. A good
-          example of their use is in the{' '}
-          <Link to="/layout/bottom-bar">EuiBottomBar</Link> component.
+          <EuiCode language="js">{'color="ghost"'}</EuiCode> to any of the
+          button styles on this page. These should be used extremely rarely, and
+          are only for placing buttons on top of dark or image-based
+          backgrounds. A good example of their use is in the{' '}
+          <Link to="/layout/bottom-bar">
+            <strong>EuiBottomBar</strong>
+          </Link>{' '}
+          component.
         </p>
       ),
       snippet: buttonGhostSnippet,